+++ ideal-bus-broker/src/main/java/com/zzsong/bus/core/config/BusProperties.java
@@ -3,7 +3,6 @@ package com.zzsong.bus.core.config;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.NestedConfigurationProperty;
 import org.springframework.stereotype.Component;
 
 import java.time.Duration;
@@ -24,7 +23,4 @@ public class BusProperties {
    * 本地缓存的刷新间隔
    */
   private Duration refreshLocalCacheInterval = Duration.ofMinutes(10);
-
-  @NestedConfigurationProperty
-  private ThreadPoolProperties blockPool = new ThreadPoolProperties();
 }