+++ ideal-bus-client-spring-boot-starter/src/main/java/com/zzsong/bus/client/spring/boot/starter/BusClientProperties.java
@@ -19,5 +19,5 @@ public class BusClientProperties {
 
   private String accessToken = "";
 
-  private boolean autoSubscribe = false;
+  private boolean autoSubscribe = true;
 }