+++ ideal-bus-samples/ideal-bus-sample-client/src/main/java/com/zzsong/bus/sample/client/SampleController.java
@@ -11,10 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.annotation.Nonnull;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 
 /**
  * @author 宋志宗 on 2020/9/20 2:31 下午
@@ -25,7 +22,6 @@ import java.util.concurrent.Executors;
 public class SampleController {
   @Nonnull
   private final EventPublisher eventPublisher;
-  private final ExecutorService executorService = Executors.newFixedThreadPool(16);
 
   public SampleController(@Nonnull EventPublisher eventPublisher) {
     this.eventPublisher = eventPublisher;
@@ -36,8 +32,8 @@ public class SampleController {
    */
   @GetMapping("/testBlock/{topic}")
   public void testBlock(@PathVariable("topic") String topic) {
-    final long start = System.currentTimeMillis();
-    final List<String> payload = ImmutableList.of("1", "2", "3");
+    long start = System.currentTimeMillis();
+    List<String> payload = ImmutableList.of("1", "2", "3");
     eventPublisher.publish(EventMessage.of(topic, payload))
         .map(JsonUtils::toJsonString)
         .doOnNext(log::info)
@@ -47,39 +43,21 @@ public class SampleController {
 
   /**
    * 这是一个异步发布事件的示例
-   * <pre>
-   *   .subscribe() 终结符默认会在reactor线程中执行事件发布, 正常情况下这是一个非IO阻塞的操作.
-   *   注意: 在reactor管道中执行IO阻塞操作将会快速消耗reactor线程, 并发量大的情况下可能会导致服务不可用
-   * </pre>
    */
   @GetMapping("/testAsync/{topic}")
   public void testAsync(@PathVariable("topic") String topic) {
-    final long start = System.currentTimeMillis();
-    final List<String> payload = ImmutableList.of("1", "2", "3");
+    long start = System.currentTimeMillis();
+    List<String> payload = ImmutableList.of("1", "2", "3");
     EventMessage<List<String>> message = EventMessage.of(topic, payload)
+        .bizId("123456")
+        .key("key")
+        .externalApp("externalApp")
         .addHeader("age", "20");
     eventPublisher.publish(message)
         .map(JsonUtils::toJsonString)
         .doOnNext(log::info)
         .doFinally(s -> log.info("发布耗时: {}", System.currentTimeMillis() - start))
-        .subscribe();
+        .subscribe(); // .subscribe() 终结符默认会在reactor线程中执行事件发布, 正常情况下这是一个非IO阻塞的操作.
     log.info("完成发布, 当前线程耗时: {}", System.currentTimeMillis() - start);
-  }
-
-  @GetMapping("/testHttp/{topic}/{loop}/{times}")
-  public void testHttp(@PathVariable("topic") String topic, @PathVariable("loop") int loop, @PathVariable("times") int times) {
-    for (int i = 0; i < loop; i++) {
-      executorService.submit(() -> {
-        final long start = System.currentTimeMillis();
-        List<EventMessage<?>> list = new ArrayList<>(times);
-        for (int j = 0; j < times; j++) {
-          final List<String> payload = ImmutableList.of(j + 1 + "", j + 1 + "", j + 1 + "");
-          EventMessage<List<String>> message = EventMessage.of(topic, payload);
-          list.add(message);
-        }
-        eventPublisher.batchPublish(list).collectList().block();
-        log.info("完成发布, 当前线程耗时: {}", System.currentTimeMillis() - start);
-      });
-    }
   }
 }